Moving bytes signing helpers from `core` to `storage`.
